moved createAuthToken function to util, issued JWT upon registration and stored in local storage, updated tests appropriately
