+++ auth/router.js
@@ -2,17 +2,8 @@
 const express = require('express');
 const passport = require('passport');
 const bodyParser = require('body-parser');
-const jwt = require('jsonwebtoken');
-const {JWT_SECRET, JWT_EXPIRY} = require('../config');
+const {createAuthToken} = require('../util');
 const router = express.Router(); 
-
-const createAuthToken = function(user) {
-	return jwt.sign({user}, JWT_SECRET, {
-		subject: user.username, 
-		expiresIn: JWT_EXPIRY, 
-		algorithm: 'HS256'
-	});
-};
 
 const localAuth = passport.authenticate('local', {session: false});
 router.use(bodyParser.json());
@@ -27,7 +18,7 @@ router.post('/refresh', jwtAuth, (req, res) => {
 	res.json({authToken});
 });
 
-router.get('/dashboard', jwtAuth, (req, res) => {
+router.get('/', jwtAuth, (req, res) => {
 	res.json("success!");
 });
